requirements.txt
streamlit==1.31.0
openai==1.54.0
httpx==0.27.2
requests==2.31.0
python-dotenv==1.0.0